host-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
requests==2.32.3
